+++ src/main/java/com/guardias/backend/repository/ProfesionRepository.java
@@ -4,10 +4,10 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.guardias.backend.entity.Profesion;
 
-public interface ProfesionRepository extends JpaRepository<Profesion, Integer> {
+public interface ProfesionRepository extends JpaRepository<Profesion, Long> {
     
-    Optional<Profesion> findById(int id);
-    boolean existsById(int id);
+    Optional<Profesion> findById(Long id);
+    boolean existsById(Long id);
 
     Optional<Profesion> findByNombre(String nombre);
     boolean existsByNombre(String nombre);
