+++ src/main/java/com/guardias/backend/repository/SuspencionRepository.java
@@ -5,10 +5,10 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.guardias.backend.entity.Suspencion;
 
-public interface SuspencionRepository extends JpaRepository<Suspencion,Integer> {
+public interface SuspencionRepository extends JpaRepository<Suspencion,Long> {
 
-    Optional<Suspencion> findById(int id); 
-    boolean existsById(int id);
+    Optional<Suspencion> findById(Long id); 
+    boolean existsById(Long id);
 
     Optional<Suspencion> findByFechaInicio(Date fechaInicio); 
     boolean existsByFechaInicio(Date fechaInicio);
