+++ src/main/java/com/guardias/backend/repository/TipoRevistaRepository.java
@@ -4,7 +4,7 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.guardias.backend.entity.TipoRevista;
 
-public interface TipoRevistaRepository extends JpaRepository<TipoRevista, Integer> {
+public interface TipoRevistaRepository extends JpaRepository<TipoRevista, Long> {
     Optional<TipoRevista> findByNombre(String nombre);
     boolean existsByNombre(String nombre);
