+++ src/main/java/com/guardias/backend/service/CargaHorariaService.java
@@ -2,10 +2,8 @@ package com.guardias.backend.service;
 
 import java.util.List;
 import java.util.Optional;
-
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
 import com.guardias.backend.entity.CargaHoraria;
 import com.guardias.backend.repository.CargaHorariaRepository;
 
@@ -19,7 +17,7 @@ public class CargaHorariaService {
         return cargaHorariaRepository.findAll();
     }
 
-    public Optional<CargaHoraria> getOne(int id) {
+    public Optional<CargaHoraria> getOne(Long id) {
         return cargaHorariaRepository.findById(id);
     }
 
@@ -31,11 +29,11 @@ public class CargaHorariaService {
         cargaHorariaRepository.save(cargaHoraria);
     }
 
-    public void delete(int id) {
+    public void delete(Long id) {
         cargaHorariaRepository.deleteById(id);
     }
 
-    public boolean existsById(int id) {
+    public boolean existsById(Long id) {
         return cargaHorariaRepository.existsById(id);
     }
